Conflicts:
	app/code/Magento/Rma/Block/Returns/Create.php
	app/code/Magento/Sales/Model/Config/Source/Order/Status.php
	dev/tests/integration/testsuite/Magento/Outbound/Authentication/FactoryTest.php
	dev/tests/integration/testsuite/Magento/Outbound/Formatter/FactoryTest.php
	dev/tests/integration/testsuite/Magento/Outbound/Formatter/JsonTest.php
	dev/tests/integration/testsuite/Magento/PubSub/Event/QueueWriterTest.php
	dev/tests/integration/testsuite/Magento/PubSub/_files/config.xml
	dev/tests/static/testsuite/Magento/Test/Legacy/_files/obsolete_classes.php
	dev/tests/static/testsuite/Magento/Test/Legacy/_files/obsolete_methods.php
	dev/tests/static/testsuite/Magento/Test/Legacy/_files/obsolete_methods_ee.php
	dev/tests/unit/testsuite/Magento/Outbound/Authentication/FactoryTest.php
	dev/tests/unit/testsuite/Magento/Outbound/Authentication/HmacTest.php
	dev/tests/unit/testsuite/Magento/Outbound/Formatter/FactoryTest.php
	dev/tests/unit/testsuite/Magento/Outbound/Formatter/JsonTest.php
	dev/tests/unit/testsuite/Magento/Outbound/Message/FactoryTest.php
	dev/tests/unit/testsuite/Magento/Outbound/MessageTest.php
	dev/tests/unit/testsuite/Magento/Outbound/Transport/Http/ResponseTest.php
	dev/tests/unit/testsuite/Magento/Outbound/Transport/HttpTest.php
	dev/tests/unit/testsuite/Magento/PubSub/Event/FactoryTest.php
	dev/tests/unit/testsuite/Magento/PubSub/Event/QueueWriterTest.php
	dev/tests/unit/testsuite/Magento/PubSub/Message/DispatcherAsyncTest.php
	lib/Magento/Outbound/Authentication/Factory.php
	lib/Magento/Outbound/Authentication/Hmac.php
	lib/Magento/Outbound/EndpointInterface.php
	lib/Magento/Outbound/Formatter/Factory.php
	lib/Magento/Outbound/Message.php
	lib/Magento/Outbound/Message/Factory.php
	lib/Magento/Outbound/Message/FactoryInterface.php
	lib/Magento/Outbound/Transport/Http.php
	lib/Magento/Outbound/Transport/Http/Response.php
	lib/Magento/Outbound/TransportInterface.php
	lib/Magento/PubSub/Event/QueueHandler.php
	lib/Magento/PubSub/EventInterface.php
	lib/Magento/PubSub/Job/QueueHandler.php
	lib/Magento/PubSub/JobInterface.php
	lib/Magento/PubSub/Message/DispatcherAsync.php
	lib/Magento/PubSub/SubscriptionInterface.php

+++ app/code/Magento/Reward/Model/Action/OrderExtra.php
@@ -41,10 +41,8 @@ class OrderExtra extends \Magento\Reward\Model\Action\AbstractAction
      * @param \Magento\Reward\Helper\Data $rewardData
      * @param array $data
      */
-    public function __construct(
-        \Magento\Reward\Helper\Data $rewardData,
-        array $data = array()
-    ) {
+    public function __construct(\Magento\Reward\Helper\Data $rewardData, array $data = array())
+    {
         $this->_rewardData = $rewardData;
         parent::__construct($data);
     }
@@ -70,9 +68,7 @@ class OrderExtra extends \Magento\Reward\Model\Action\AbstractAction
     public function setEntity($entity)
     {
         parent::setEntity($entity);
-        $this->getHistory()->addAdditionalData(array(
-            'increment_id' => $this->getEntity()->getIncrementId()
-        ));
+        $this->getHistory()->addAdditionalData(array('increment_id' => $this->getEntity()->getIncrementId()));
         return $this;
     }
 
@@ -104,16 +100,16 @@ class OrderExtra extends \Magento\Reward\Model\Action\AbstractAction
             // known issue: no support for multishipping quote
             $address = $quote->getIsVirtual() ? $quote->getBillingAddress() : $quote->getShippingAddress();
             // use only money customer spend - shipping & tax
-            $monetaryAmount = $quote->getBaseGrandTotal()
-                - $address->getBaseShippingAmount()
-                - $address->getBaseTaxAmount();
+            $monetaryAmount = $quote->getBaseGrandTotal() -
+                $address->getBaseShippingAmount() -
+                $address->getBaseTaxAmount();
             $monetaryAmount = $monetaryAmount < 0 ? 0 : $monetaryAmount;
         } else {
-            $monetaryAmount = $this->getEntity()->getBaseTotalPaid()
-                - $this->getEntity()->getBaseShippingAmount()
-                - $this->getEntity()->getBaseTaxAmount();
+            $monetaryAmount = $this->getEntity()->getBaseTotalPaid() -
+                $this->getEntity()->getBaseShippingAmount() -
+                $this->getEntity()->getBaseTaxAmount();
         }
-        $pointsDelta = $this->getReward()->getRateToPoints()->calculateToPoints((float)$monetaryAmount);
+        $pointsDelta = $this->getReward()->getRateToPoints()->calculateToPoints((double)$monetaryAmount);
         return $pointsDelta;
     }
 
@@ -125,7 +121,6 @@ class OrderExtra extends \Magento\Reward\Model\Action\AbstractAction
      */
     public function canAddRewardPoints()
     {
-        return parent::canAddRewardPoints()
-            && $this->_rewardData->isOrderAllowed($this->getReward()->getWebsiteId());
+        return parent::canAddRewardPoints() && $this->_rewardData->isOrderAllowed($this->getReward()->getWebsiteId());
     }
 }
